Delete duplicated in core_sys_info.py: - imports - users_info(), pasword_policy(), password_expiry() --> now in core_user.py Move password_confirm() from core_sys_info.py to core_user.py Update README.md Update requirements.txt --> Added cryptography module
