+++ synology_api/core_sys_info.py
@@ -1,9 +1,6 @@
 from __future__ import annotations
 from typing import Optional
 from . import base_api
-from __future__ import annotations
-from typing import Optional
-from . import base_api
 
 
 class SysInfo(base_api.BaseApi):
@@ -771,62 +768,6 @@ class SysInfo(base_api.BaseApi):
         info = self.core_list[api_name]
         api_path = info['path']
         req_param = {'version': info['maxVersion'], 'method': 'get'}
-
-        return self.request_data(api_name, api_path, req_param)
-
-    def users_info(self, offset: int = 0, limit: int = -1) -> dict[str, object] | str:
-        api_name = 'SYNO.Core.User'
-        info = self.core_list[api_name]
-        api_path = info['path']
-        req_param = {'version': info['maxVersion'], 'method': 'list', 'type': 'local', 'offset': offset, 'limit': limit}
-
-        return self.request_data(api_name, api_path, req_param)
-
-    def password_policy(self) -> dict[str, object] | str:
-        api_name = 'SYNO.Core.User.PasswordPolicy'
-        info = self.core_list[api_name]
-        api_path = info['path']
-        req_param = {'version': info['maxVersion'], 'method': 'get'}
-
-        return self.request_data(api_name, api_path, req_param)
-
-    def password_expiry(self) -> dict[str, object] | str:
-        api_name = 'SYNO.Core.User.PasswordExpiry'
-        info = self.core_list[api_name]
-        api_path = info['path']
-        req_param = {'version': info['maxVersion'], 'method': 'get'}
-
-        return self.request_data(api_name, api_path, req_param)
-    
-    def password_confirm(self, password: str) -> dict[str, object] | str:
-        """Issues a passowrd/session comparison to ensure the given password matches the auth of the current session.
-
-        This is needed by some APIs as a confirmation method, for example, when creating/modifying a scheduled task with root permissions. 
-        Please note that the password will be sent in plain text, just like in the base auth method.
-
-        Args:
-            password (str):
-                The password with which the session was initiated.
-
-        Returns:
-            dict|str:
-                A dictionary containing a `SynoConfirmPWToken`, or an error message.
-
-            Example return:
-            {
-                "data": {
-                    "SynoConfirmPWToken": "xxxxx"
-                },
-                "success": true
-            }
-        """
-        # There is a way to send the password encrypted, but could not figure out how the NAS wants the data to be encrypted.
-        # It wants an AES and RSA key sent to it under the field "__cIpHeRtExT", tried some ways of implementing it, 
-        # but kept getting decryption errors in /var/log/synoscgi.log, so just went through with the plain text password. 
-        api_name = 'SYNO.Core.User.PasswordConfirm'
-        info = self.core_list[api_name]
-        api_path = info['path']
-        req_param = {'version': 2, 'method': 'auth', 'password': password}
 
         return self.request_data(api_name, api_path, req_param)
